pyproject.toml
[tool.poetry]
name = "cn_stats_data"
authors = ["Eric Feng <fengyj@live.com>"]
description = "Crawler for statstic data from National Bureau of Statistic of China"
version = "0.0.1"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.3"
pyyaml = "^6.0.2"
python-dateutil = "^2.9.0.post0"
psycopg2 = "^2.9.10"
retry = "^0.9.2"
cn-stats-util = {path = "D:/Personal/cn-stats-util"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
